+++ backend/Tests/Interop/InteropBindingTests.cs
@@ -73,4 +73,27 @@ public class InteropBindingTests
             .UseFileName("event")
             .UseDirectory("__snapshots__");
     }
+    
+    [Theory]
+    [InlineData(ModuleSchemaVersion.V0, (byte)0)]
+    [InlineData(ModuleSchemaVersion.V2, (byte)2)]
+    public void GivenVersion_WhenMapModuleSchemaVersionToFFIOption_ThenOptionContainsVersion(ModuleSchemaVersion version, byte mapped)
+    {
+        // Act
+        var ffiOption = InteropBinding.FFIByteOption.Create(version);
+
+        // Assert
+        ffiOption.is_some.Should().Be(1);
+        ffiOption.t.Should().Be(mapped);
+    }
+    
+    [Fact]
+    public void GivenUndefinedVersion_WhenMapModuleSchemaVersionToFFIOption_ThenOptionEmpty()
+    {
+        // Act
+        var ffiOption = InteropBinding.FFIByteOption.Create(ModuleSchemaVersion.Undefined);
+
+        // Assert
+        ffiOption.is_some.Should().Be(0);
+    }
 }
